Implement both for ppx_let with apply instead of bind (fixes #3)
Adds `apply_both` to the `Apply` type class. It seemed perfect because of its symmetry to `apply_first`. Formally `both` is the `Monoidal` equivalent of `apply`. I thought `both` would need `pure` for lifted `map` but it worked without it. I also considered supplying overrides but it's tricky to implement lawfully.